moves baseplugin to a separated file
